displaying error message
I'm trying to display an error message on the screen if a user inputs a player that is not in the database on the red or blue team. I am currently working on it but about to go to class so that is why I am this with my code commented out.